requirements.txt
numpy>=1.20.0
scipy>=1.7.0
Pillow>=9.0.0
matplotlib>=3.5.0
pandas>=1.3.0
kagglehub>=0.3.5
flask>=2.0.0
werkzeug>=2.0.0